+++ src/injection/classes-gjs.ts
@@ -1,7 +1,8 @@
 import type { InjectionClass } from '../types/index.js'
 
-export const classes: InjectionClass[] = [
+export const classesGjs: InjectionClass[] = [
     {
+        versions: ['2.0'],
         qualifiedName: 'GObject.ParamSpec',
         // Static functions injected by GJS, see https://gitlab.gnome.org/GNOME/gjs/-/blob/master/modules/core/overrides/GObject.js
         staticFunctions: [
@@ -631,48 +632,10 @@ export const classes: InjectionClass[] = [
             },
         ],
     },
-    {
-        qualifiedName: 'Gio.ListStore',
-        generics: [
-            {
-                name: 'A',
-                extends: 'GObject.Object',
-                value: 'GObject.Object',
-            },
-        ],
-    },
-    {
-        qualifiedName: 'GLib.List',
-        generics: [
-            {
-                name: 'A',
-                value: 'any',
-            },
-        ],
-    },
-    {
-        qualifiedName: 'GLib.HashTable',
-        // TODO
-        // [key: A]: B;
-        // properties: [
-        //     {
-        //         name: '[key: A]'
-        //         type: 'B'
-        //     }
-        // ],
-        generics: [
-            {
-                name: 'A',
-                value: 'symbol | string | number',
-            },
-            {
-                name: 'B',
-                value: 'string | number | boolean',
-            },
-        ],
-    },
+
     // https://gjs.guide/guides/glib/gvariant.html#unpacking-variants
     {
+        versions: ['2.0'],
         qualifiedName: 'GLib.Variant',
         methods: [
             {