+++ src/injections/index.ts
@@ -1,2 +1,0 @@
-export * from './classes.js'
-export * from './type-patches.js'